Remove obsolete crud module
